Updated first_run.sh to handle non-droppable user
If you use the container with existing data and the user already exists and is attached to a database, you will not be able to drop it. It will not be created again, and its password will remain the previous one.

Here we just add a line to alter the user password and ensure we change it in all cases.